requirements.txt
flask
flask-cors
werkzeug
python-dotenv
pillow
requests
pandas
openpyxl
gunicorn
